refactor(api): 공통 컴포넌트 관리를 위한 폴더 위치변경
- 유저 api -> apis 폴더로 위치변경
